+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Fix in Euro Heater to correctly set desired mode
+
 ## 1.1.39 (2022-10-14)
 
 * (theimo1221) Minor fixes in Postgres DDL Code